+++ tools/pd-ctl/pdctl/command/label_command.go
@@ -15,10 +15,15 @@
 package command
 
 import (
+	"encoding/json"
 	"fmt"
 	"net/http"
 
 	"github.com/spf13/cobra"
+	"github.com/tikv/pd/server/api"
+	"github.com/tikv/pd/server/config"
+	"github.com/tikv/pd/server/core"
+	"github.com/tikv/pd/server/statistics"
 )
 
 var (
@@ -34,6 +39,7 @@ func NewLabelCommand() *cobra.Command {
 		Run:   showLabelsCommandFunc,
 	}
 	l.AddCommand(NewLabelListStoresCommand())
+	l.AddCommand(NewCheckLabels())
 	return l
 }
 
@@ -77,4 +83,104 @@ func showLabelListStoresCommandFunc(cmd *cobra.Command, args []string) {
 		return
 	}
 	cmd.Println(r)
+}
+func NewCheckLabels() *cobra.Command {
+	return &cobra.Command{
+		Use:     "isolation [label]",
+		Short:   "isolation labels",
+		Example: "label-count: map[isolation-label:count], region-map: [region-id:isolation-label]",
+		Run:     checkIsolationLabel,
+	}
+}
+
+func getReplicationConfig(cmd *cobra.Command, _ []string) (*config.ReplicationConfig, error) {
+	prefix := configPrefix + "/replicate"
+	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
+	if err != nil {
+		return nil, err
+	}
+	var config config.ReplicationConfig
+	if err := json.Unmarshal([]byte(body), &config); err != nil {
+		return nil, err
+	}
+	return &config, nil
+}
+
+func getStores(cmd *cobra.Command, _ []string) ([]*core.StoreInfo, error) {
+	prefix := storesPrefix
+	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
+	if err != nil {
+		return nil, err
+	}
+	var storesInfo api.StoresInfo
+	if err := json.Unmarshal([]byte(body), &storesInfo); err != nil {
+		return nil, err
+	}
+	stores := make([]*core.StoreInfo, 0)
+	for _, storeInfo := range storesInfo.Stores {
+		stores = append(stores, core.NewStoreInfo(storeInfo.Store.Store))
+	}
+	return stores, nil
+}
+
+func getRegions(cmd *cobra.Command, _ []string) ([]api.RegionInfo, error) {
+	prefix := regionsPrefix
+	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
+	if err != nil {
+		return nil, err
+	}
+	var RegionsInfo api.RegionsInfo
+	if err := json.Unmarshal([]byte(body), &RegionsInfo); err != nil {
+		return nil, err
+	}
+	return RegionsInfo.Regions, nil
+}
+
+func checkIsolationLabel(cmd *cobra.Command, args []string) {
+	var checkLabel string
+	if len(args) == 1 {
+		checkLabel = args[0]
+	}
+
+	storesInfo, err := getStores(cmd, args)
+	if err != nil {
+		cmd.Printf("Failed to get stores info: %s\n", err)
+		return
+	}
+
+	regionsInfo, err := getRegions(cmd, args)
+	if err != nil {
+		cmd.Printf("Failed to get regions info: %s\n", err)
+		return
+	}
+
+	config, err := getReplicationConfig(cmd, args)
+	if err != nil {
+		cmd.Printf("Failed to get labels: %s\n", err)
+		return
+	}
+	locationLabels := config.LocationLabels
+
+	storesMap := make(map[uint64]*core.StoreInfo, len(storesInfo))
+	for _, store := range storesInfo {
+		storesMap[store.GetID()] = store
+	}
+
+	regionMap := make(map[uint64]string, len(regionsInfo))
+	labelCount := make(map[string]int, len(locationLabels))
+
+	for _, region := range regionsInfo {
+		stores := make([]*core.StoreInfo, 0)
+		for _, peer := range region.Peers {
+			if s, ok := storesMap[peer.StoreId]; ok {
+				stores = append(stores, s)
+			}
+		}
+		isolationLabel := statistics.GetRegionLabelIsolation(stores, locationLabels)
+		if len(checkLabel) == 0 || isolationLabel == checkLabel {
+			regionMap[region.ID] = isolationLabel
+			labelCount[isolationLabel]++
+		}
+	}
+	cmd.Printf("label-count:%+v,region-map:%+v\n", labelCount, regionMap)
 }
