+++ tools/pd-ctl/pdctl/command/region_command.go
@@ -19,10 +19,6 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"github.com/tikv/pd/server/api"
-	"github.com/tikv/pd/server/config"
-	"github.com/tikv/pd/server/core"
-	"github.com/tikv/pd/server/statistics"
 	"io"
 	"net/http"
 	"net/url"
@@ -64,7 +60,6 @@ func NewRegionCommand() *cobra.Command {
 	r.AddCommand(NewRegionWithStoreCommand())
 	r.AddCommand(NewRegionsByKeysCommand())
 	r.AddCommand(NewRangesWithRangeHolesCommand())
-	r.AddCommand(NewCheckLabels())
 
 	topRead := &cobra.Command{
 		Use:   `topread <limit> [--jq="<query string>"]`,
@@ -542,106 +537,6 @@ This command will output all empty ranges without any region info.`
   ]
 `
 )
-
-func NewCheckLabels() *cobra.Command {
-	return &cobra.Command{
-		Use:   "isolation <label>",
-		Short: "isolation labels",
-		Run:   checkLabels,
-	}
-}
-
-func getReplicationConfig(cmd *cobra.Command, _ []string) (*config.ReplicationConfig, error) {
-	prefix := configPrefix + "/replicate"
-	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
-	if err != nil {
-		return nil, err
-	}
-	var config config.ReplicationConfig
-	if err := json.Unmarshal([]byte(body), &config); err != nil {
-		return nil, err
-	}
-	return &config, nil
-}
-
-func getStores(cmd *cobra.Command, _ []string) ([]*core.StoreInfo, error) {
-	prefix := storesPrefix
-	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
-	if err != nil {
-		return nil, err
-	}
-	var storesInfo api.StoresInfo
-	if err := json.Unmarshal([]byte(body), &storesInfo); err != nil {
-		return nil, err
-	}
-	stores := make([]*core.StoreInfo, 0)
-	for _, storeInfo := range storesInfo.Stores {
-		stores = append(stores, core.NewStoreInfo(storeInfo.Store.Store))
-	}
-	return stores, nil
-}
-
-func getRegions(cmd *cobra.Command, _ []string) ([]api.RegionInfo, error) {
-	prefix := regionsPrefix
-	body, err := doRequest(cmd, prefix, http.MethodGet, http.Header{})
-	if err != nil {
-		return nil, err
-	}
-	var RegionsInfo api.RegionsInfo
-	if err := json.Unmarshal([]byte(body), &RegionsInfo); err != nil {
-		return nil, err
-	}
-	return RegionsInfo.Regions, nil
-}
-
-func checkLabels(cmd *cobra.Command, args []string) {
-	var checkLabel string
-	if len(args) == 1 {
-		checkLabel = args[0]
-	}
-
-	storesInfo, err := getStores(cmd, args)
-	if err != nil {
-		cmd.Printf("Failed to get stores info: %s\n", err)
-		return
-	}
-
-	regionsInfo, err := getRegions(cmd, args)
-	if err != nil {
-		cmd.Printf("Failed to get regions info: %s\n", err)
-		return
-	}
-
-	config, err := getReplicationConfig(cmd, args)
-	if err != nil {
-		cmd.Printf("Failed to get labels: %s\n", err)
-		return
-	}
-	locationLabels := config.LocationLabels
-
-	storesMap := make(map[uint64]*core.StoreInfo, len(storesInfo))
-	for _, store := range storesInfo {
-		storesMap[store.GetID()] = store
-	}
-
-	regionMap := make(map[uint64]string, len(regionsInfo))
-	labelCount := make(map[string]int, len(locationLabels))
-
-	for _, region := range regionsInfo {
-		stores := make([]*core.StoreInfo, 0)
-		for _, peer := range region.Peers {
-			if s, ok := storesMap[peer.StoreId]; ok {
-				stores = append(stores, s)
-			}
-		}
-		isolationLabel := statistics.GetRegionLabelIsolation(stores, locationLabels)
-		if len(checkLabel) == 0 || isolationLabel == checkLabel {
-			regionMap[region.ID] = isolationLabel
-			labelCount[isolationLabel]++
-		}
-	}
-	cmd.Printf("labelCount:%+v,regionMap:%+v\n", labelCount, regionMap)
-}
 
 // NewRangesWithRangeHolesCommand returns ranges with range-holes subcommand of regionCmd
 func NewRangesWithRangeHolesCommand() *cobra.Command {
